split up background.js
